requirements/base.txt
django==2.0.7
django-model-utils==3.1.2
djangorestframework==3.8.2
easy-thumbnails==2.5
psycopg2-binary==2.7.5
rest-condition==1.0.3
transliterate==1.10.1
